fix(playground): fall back to short id for blank experiment names
An experiment whose name is an empty string rendered a blank link in the
comparison cards, with the short id row hidden too, leaving the card with no
visible identifier. Nothing rejects an empty name server-side.

The Experiment column on the overview was also sized `auto`, so it grew to fit
the longest name and squeezed its neighbours instead of truncating.

Also deduplicates the name label shared by the dataset and overview lists into
ExperimentNameLabel, and drops ExperimentPageHeader, unreachable since the
detail page moved to ExperimentTopArea.
