+++ packages/playground/src/domains/datasets/components/experiments/dataset-experiments-list.tsx
@@ -2,11 +2,11 @@ import type { DatasetExperiment } from '@mastra/client-js';
 import { Chip } from '@mastra/playground-ui/components/Chip';
 import { DataList } from '@mastra/playground-ui/components/DataList';
 import { EmptyState } from '@mastra/playground-ui/components/EmptyState';
-import { getShortId } from '@mastra/playground-ui/components/Text';
 import { Tooltip, TooltipContent, TooltipTrigger } from '@mastra/playground-ui/components/Tooltip';
 import { cn } from '@mastra/playground-ui/utils/cn';
 import { format, isThisYear, isToday } from 'date-fns';
 import { Play } from 'lucide-react';
+import { ExperimentNameLabel } from '@/domains/experiments/components/experiment-name-label';
 
 const experimentsListColumns = [
   { name: 'experiment', label: 'Experiment', size: 'minmax(9rem,1fr)' },
@@ -16,43 +16,6 @@ const experimentsListColumns = [
   { name: 'counts', label: 'Counts', size: '7rem' },
   { name: 'date', label: 'Created', size: '10rem' },
 ];
-
-/**
- * Primary label for an experiment row: its name (falling back to the short ID),
- * with the short ID shown as a secondary line when a name is present. The
- * description, when set, surfaces in a tooltip.
- */
-function ExperimentNameCell({ experiment }: { experiment: DatasetExperiment }) {
-  const id = getShortId(experiment.id) ?? experiment.id;
-
-  if (!experiment.name) {
-    return (
-      <DataList.Cell height="compact" className="min-w-0">
-        <span className="text-ui-smd text-neutral3 block truncate font-mono">{id}</span>
-      </DataList.Cell>
-    );
-  }
-
-  const label = (
-    <span className="flex min-w-0 flex-col gap-0.5 py-0.5 text-left">
-      <span className="text-ui-md text-neutral4 block truncate">{experiment.name}</span>
-      <span className="text-ui-sm text-neutral2 block truncate font-mono">{id}</span>
-    </span>
-  );
-
-  return (
-    <DataList.Cell height="compact" className="min-w-0">
-      {experiment.description ? (
-        <Tooltip>
-          <TooltipTrigger asChild>{label}</TooltipTrigger>
-          <TooltipContent>{experiment.description}</TooltipContent>
-        </Tooltip>
-      ) : (
-        label
-      )}
-    </DataList.Cell>
-  );
-}
 
 export interface DatasetExperimentsListProps {
   experiments: DatasetExperiment[];
@@ -109,7 +72,9 @@ export function DatasetExperimentsList({
 
         const rowCells = (
           <>
-            <ExperimentNameCell experiment={experiment} />
+            <DataList.Cell height="compact" className="min-w-0">
+              <ExperimentNameLabel experiment={experiment} />
+            </DataList.Cell>
             <DataList.Cell height="compact">
               {experiment.status && (
                 <Tooltip>
